requirements.txt
fastapi
uvicorn
requests
together
python-dotenv
redis
qdrant-client
